homework 43

+++ homework43/Program.cs
@@ -22,12 +22,12 @@ Console.Write($"({intersectionPointForX};");
 double intersectionPointForY = Math.Round(IntersectionPointForY(b1, k1, intersectionPointForX), 1);
 Console.WriteLine($" {intersectionPointForY})");
 
-double IntersectionPointForX (int b1c, int k1c, int b2c, double k2c) //почему без 1го параметра double не получается все формула? Как это число конвертируется?
+double IntersectionPointForX (int b1c, int k1c, int b2c, int k2c) //почему без 1го параметра double не получается все формула? Как это число конвертируется?
 {
-    return (b2c - b1c) / (k1c - k2c);
+    return (double)(b2c - b1c) / (k1c - k2c);
 }
 
 double IntersectionPointForY (int b1c, int k1c, double x)
 {
-   return  k1c * x + b1c;
+   return k1c * x + b1c;
 }
